Fix issue with PhysicsAnimation::advanceSingleFrame

+++ src/examples/level_set_liquid_sim/main.cpp
@@ -26,7 +26,7 @@ using namespace jet;
 void saveTriangleMesh(
     const TriangleMesh3& mesh,
     const std::string& rootDir,
-    unsigned int frameCnt) {
+    int frameCnt) {
     char basename[256];
     snprintf(basename, sizeof(basename), "frame_%06d.obj", frameCnt);
     std::string filename = pystring::os::path::join(rootDir, basename);
@@ -41,7 +41,7 @@ void saveTriangleMesh(
 void triangulateAndSave(
     const ScalarGrid3Ptr& sdf,
     const std::string& rootDir,
-    unsigned int frameCnt) {
+    int frameCnt) {
     TriangleMesh3 mesh;
     int flag = kDirectionAll & ~kDirectionDown;
     marchingCubes(
@@ -89,7 +89,7 @@ void printInfo(const LevelSetLiquidSolver3Ptr& solver) {
 void runSimulation(
     const std::string& rootDir,
     const LevelSetLiquidSolver3Ptr& solver,
-    size_t numberOfFrames,
+    int numberOfFrames,
     double fps) {
     auto sdf = solver->signedDistanceField();
 
@@ -104,7 +104,7 @@ void runSimulation(
 void runExample1(
     const std::string& rootDir,
     size_t resX,
-    unsigned int numberOfFrames,
+    int numberOfFrames,
     double fps) {
     // Build solver
     auto solver = LevelSetLiquidSolver3::builder()
@@ -149,7 +149,7 @@ void runExample1(
 void runExample2(
     const std::string& rootDir,
     size_t resX,
-    unsigned int numberOfFrames,
+    int numberOfFrames,
     double fps) {
     // Build solver
     auto solver = LevelSetLiquidSolver3::builder()
@@ -224,7 +224,7 @@ void runExample2(
 void runExample3(
     const std::string& rootDir,
     size_t resX,
-    unsigned int numberOfFrames,
+    int numberOfFrames,
     double fps) {
     // Build solver
     auto solver = LevelSetLiquidSolver3::builder()
@@ -270,7 +270,7 @@ void runExample3(
 void runExample4(
     const std::string& rootDir,
     size_t resX,
-    unsigned int numberOfFrames,
+    int numberOfFrames,
     double fps) {
     // Build solver
     auto solver = LevelSetLiquidSolver3::builder()
@@ -314,7 +314,7 @@ void runExample4(
 
 int main(int argc, char* argv[]) {
     size_t resX = 50;
-    unsigned int numberOfFrames = 100;
+    int numberOfFrames = 100;
     double fps = 60.0;
     int exampleNum = 1;
     std::string logFilename = APP_NAME ".log";
@@ -341,7 +341,7 @@ int main(int argc, char* argv[]) {
                 resX = static_cast<size_t>(atoi(optarg));
                 break;
             case 'f':
-                numberOfFrames = static_cast<size_t>(atoi(optarg));
+                numberOfFrames = atoi(optarg);
                 break;
             case 'p':
                 fps = atof(optarg);
